pyproject.toml
[project]
name = "native-language-disease-qa-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-ollama==1.1.0",
    "langchain-openrouter==0.2.1",
    "langchain-community>=0.4.1",
    "pypdf>=6.10.2",
    "pyyaml>=6.0.3",
    "python-dotenv>=1.2.2",
    "ruff>=0.15.11",
    "chromadb==1.5.8",
    "langchain-chroma==1.1.0",
    "langchain-text-splitters==1.1.2",
    "langchain-google-genai>=4.2.2",
    "fastapi>=0.136.1",
    "python-multipart>=0.0.26",
    "uvicorn>=0.46.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
]
